remove `interpose` in `MOCK_POOL_NOSSL`
Returning `nullptr` can result in exceptions. Removing the call to `interpose` will have the default behavior call the C driver function. The behavior can be overriden by the test.

Use `visit` in test functions that do not need to override the C driver function. `visit` is called in addition to the C driver function.